requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.1
gTTS==2.3.2
gunicorn==21.2.0
Pillow==10.0.0
pytesseract==0.3.10
Werkzeug==3.0.1